Added additional tests for error testing, memory, and the remaining operations

+++ src/main/java/conversion.java
@@ -1,9 +1,9 @@
 public class conversion {
-    public int convert4To10(long i) throws inputBoundException{
+    public int convert4To10(int i) throws inputBoundException{
         return 0;
     }
 
-    public int convert10To4(long i) throws inputBoundException{
+    public int convert10To4(int i) throws inputBoundException{
         return 0;
     }
 }
